// src/app/models/legajo/leg_eva_renovacion_ratificacion.ts
export interface Leg_Eva_RenovacionRatificacion {
    nRenRatCodigo: number;
    nLegDatCodigo: number;
    cPerCodigo: string;
    cPerApellido: string;
    cPerNombre: string;
    cCargo: string;
    cArea: string;
    cPerNroDoc: string;

    nLegRenRatEDD: number;
    nLegRenRatCD: number;
    nLegRenRatPC: number;
    nLegRenRatPromedio: number;

    cLegRenRatCondicion : string;
    cLegRenRatRenRat:string;
}