// src/app/models/legajo/leg_eva_capacitaciones.ts
export interface Leg_Eva_Capacitaciones {
    nLegDatCodigo: number;
    ponente_inter: number;
    ponente_nacional: number;
    organizador_inter: number;
    organizador_nacional: number;
    asistente_inter: number;
    asistente_nacional: number;
    capa_1:number;
    capa_2:number;
    capa_3:number;
    capa_4:number;
    puntaje:number;
}